[BOSK-89] Refactor block.base: remove separate slots definition; fix imports
