+++ bosk/__init__.py
@@ -3,7 +3,7 @@ __version__ = '0.1'
 
 from .data import Data
 from .stages import Stages
-from .block.slot import BaseSlot, BlockInputSlot, BlockOutputSlot, SlotT
+from .block.base import BaseSlot, BlockInputSlot, BlockOutputSlot, SlotT
 
 
 __all__ = [
