+++ bosk/block/meta.py
@@ -1,6 +1,33 @@
 from dataclasses import dataclass
 from typing import Mapping, List
-from .slot import InputSlotMeta, OutputSlotMeta, list_of_slots_meta_to_mapping
+from ..stages import Stages
+
+
+@dataclass(eq=True, frozen=True)
+class BaseSlotMeta:
+    """Base slot meta information.
+
+    Slot meta is unique for a block slot.
+
+    Attributes:
+        name: Slot name.
+        stages: At which stages slot value is needed.
+
+    """
+    name: str
+    stages: Stages = Stages()
+
+
+@dataclass(eq=True, frozen=True)
+class InputSlotMeta(BaseSlotMeta):
+    """Block input slot meta.
+    """
+
+
+@dataclass(eq=True, frozen=True)
+class OutputSlotMeta(BaseSlotMeta):
+    """Block output slot meta.
+    """
 
 
 @dataclass
@@ -37,11 +64,29 @@ class BlockMeta:
     def __init__(self, *, inputs=None, outputs=None, execution_props=None):
         assert inputs is not None, 'Meta inputs description is required'
         assert outputs is not None, 'Meta outputs description is required'
-        self.inputs = list_of_slots_meta_to_mapping(inputs)
-        self.outputs = list_of_slots_meta_to_mapping(outputs)
+        self.inputs = BlockMeta.list_of_slots_meta_to_mapping(inputs)
+        self.outputs = BlockMeta.list_of_slots_meta_to_mapping(outputs)
         if execution_props is None:  # use default properties
             execution_props = BlockExecutionProperties()
         self.execution_props = execution_props
+
+    @staticmethod
+    def list_of_slots_meta_to_mapping(slots_meta_list: List[BaseSlotMeta]) -> Mapping[str, BaseSlotMeta]:
+        """Convert list of slots meta to mapping (name -> slot meta).
+
+        Args:
+            slots_meta_list: List of slots meta.
+
+        Returns:
+            Mapping dict (name -> slot meta).
+
+        """
+        return {
+            slot_meta.name: slot_meta
+            for slot_meta in slots_meta_list
+        }
+
+
 
 
 def make_simple_meta(input_names: List[str], output_names: List[str], **additional_params) -> BlockMeta:
