+++ bosk/block/zoo/models/classification/mgs_ferns.py
@@ -9,8 +9,7 @@ from typing import Optional, Tuple, Union
 from functools import partial, reduce
 
 from ....base import BaseBlock, TransformOutputData, BlockInputData
-from ....meta import BlockMeta, BlockExecutionProperties
-from ....slot import InputSlotMeta, OutputSlotMeta
+from ....meta import BlockMeta, BlockExecutionProperties, InputSlotMeta, OutputSlotMeta
 from .....stages import Stages
 from .....data import CPUData, GPUData
 from .....utility import get_random_generator, get_rand_int
