+++ bosk/block/zoo/output_plugs/__init__.py
@@ -1,0 +1,1 @@
+from .output_plugs import OutputBlock
